Rename and deprecated rfc8888 package
