+++ pkg/ccfb/ticker.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2026 The Pion community <https://pion.ly>
 // SPDX-License-Identifier: MIT
 
-package rfc8888
+package ccfb
 
 import "time"
